scripts/requirements.txt
tensorflow
keras==2.13.1
numpy==1.24.3
opencv-python==4.8.0.74
Pillow==10.0.0
matplotlib==3.7.2
fastapi==0.103.1
uvicorn==0.23.2
python-multipart==0.0.6
